requirements.txt
tensorboardX==2.6.2.2
packaging==23.2
monai==1.1.0
dynamic-network-architectures==0.3.1
calflops==0.2.9
scipy==1.12.0
batchgenerators==0.25
ml_collections
timm==0.9.16
torchio==0.19.6
